made some of my INT's more explicit

+++ albums_migration.sql
@@ -7,8 +7,8 @@ CREATE TABLE albums
     id           INT UNSIGNED NOT NULL AUTO_INCREMENT,
     artist       VARCHAR(255)     NOT NULL,
     name         VARCHAR(255)      NOT NULL,
-    release_date INT UNSIGNED NOT NULL,
-    sales        FLOAT        NOT NULL,
+    release_date INT(4) UNSIGNED NOT NULL,
+    sales        FLOAT(5,2)        NOT NULL,
     genre        VARCHAR(255)      NOT NULL,
     PRIMARY KEY (id)
 );
